common/RefCountedObj: cleanup con/des
Also, don't allow children to set nref (to 0). This is the more significant
change as it required fixing various code to not do this:

   <reftype> ptr = new RefCountedObjectFoo(...);

as a way to create a starting reference with nref==1. This is a pretty bad code
smell so I've converted all the code doing this to use the new ::create()
factory method which produces the reference safely.

libradosstriper was particularly egregious in its abuse of setting the starting
nref. :(

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>