qa: force creation of fs with EC default data pool
In the future, we should add the EC data pool as a supplementary data
pool but that requires a mount to setup which is awkward in the code
here. When cephfs-shell is more widely available, this will be easier.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>
(cherry picked from commit 6e448f9aa0e71f1e820f4bd1b53e067268e4b67a)