mds/FSMap: assign default_compat to pre-v16.2.5 standby daemons
With v16.2.5, MDS sends its CompatSet in its beacons to the mons. If an
older MDS fails and rejoins the cluster, it gets assigned the empty
CompatSet. This is problematic during upgrades as an MDS failure may
prevent the upgrade process from continuing and cause file system
unavailability.

This patch makes it so the mons will assign a reasonable default: a
CompatSet used since v14.2.0 until v16.2.5.

Fixes: https://tracker.ceph.com/issues/53150
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>