mds: skip inode with cap iteration for empty dir
This avoids an expensive iteration over all inodes with caps in a
SnapRealm if the child is already empty: it cannot have any inodes in
cache that reference it as an ancestor. This optimizes particularly for
the case where a directory is being removed as part of a recursive
unlink.

Fixes: https://tracker.ceph.com/issues/53192
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>