mds: remove CDir::pop_spread
This is a old artifact of when it was thought MDS requests would be
forwarded. It seemed good to track popularity based on who sent the
request. So, "who" is almost always -1. Because of this, a directory's
CDir::pop_spread is never actually hit. This has the side-effect of
preventing any replication except from hits from open/getattr!

Instead of using pop_spread, just use CDir::pop_me. Also, replicate
for any read load including READDIR.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>