qa: use postmerge hooks for filtering stock kernel
Up to now, the k-stock kernel required an awkward matrix configuration
to ensure that we're only testing the rhel stock kernel. This is
basically done by overriding (evil) the yaml fragment specifying the
distribution.  So, you'd have a config like:

	fs/snaps/{begin clusters/1a3s-mds-1c-client conf/{client mds mon osd} distro/{ubuntu_latest} mount/kclient/{mount overrides/{distro/stock/{k-stock rhel_8} ms-die-on-skipped}} objectstore-ec/bluestore-bitmap overrides/{whitelist_health whitelist_wrongly_marked_down} tasks/workunit/snaps}

One would rightly believe it's running with ubuntu but, actually, the
rhel_8 fragment overrides this.

Instead of this, we're using the new postmerge teuthology hooks to
filter our configs containing k-stock when the rhel distribution is not
selected.

This is great (and a small change by itself) but doing so would bias the
selection of jobs towards k-testing/fuse which do not drop configs not
using rhel. To work around this, we add another two rhel (three total)
fragments to the custom `qa/cephfs/distro` to compensate. The extra two
rhel fragments drop if the k-stock config is not in use.  This also
uses the teuthology postmerge hooks.

Simultaneously, we're dropping the random selection of distributions as
it's no longer necessary with nested subsets (another newer teuthology
feature). Instead, modify fs:* sub-suites to divide jobs by 2 to
approximate.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>