mds: write EResetJournal to create/close log
Prior to this set of commits, the MDS would write the ESubtreeMap to the
journal, trim everything up to that segment, then finally force the
trimming of that last segment (`MDLog::trim(0)`). This is awkward in the
new code which preserves a major segment boundary at the beginning of
the journal during trimming. Instead of writing a special case for this
situation, it seems more natural to just use EResetJournal to indicate
that the journal is beginning with a clean slate.

When replaying EResetJournal, this is okay because by this time we
have reached a point where wiping the rank specific state is okay. This
is the last step where we persist state to RADOS. Two situations may
occur:

* The MDS will fail before sending down:stopped to the mons. This will
  cause a standby to enter up:replay and wipe the
  inotable/sessionmap/etc upon reading EResetJournal. This is okay
  because these tables are to be wiped anyway when the rank is
  reactivated (see MDSRank::boot_create).

* The mons receive down:stopped and remove the rank. The reactivated MDS
  will enter MDSRank::boot_create, nuke all of the tables, and append to
  the journal without replay (never replays EResetJournal).

Finally, also change up:create to write EResetJournal to the log. Again,
this is okay because when replaying a fresh journal which has never been
trimmed, it's alright to wipe the associated tables since they will be
recreated while replaying the journal.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>