mds: add balance_automate fs setting
To turn off automatic ("default") balancer in multiple MDS clusters. The
new default is "off" as it is a constant source of problems and surprise
administrators trying multiple actives. Instead, it should be a
deliberate decision to turn it on and usually with customization like
the "bal_rank_mask" setting.

Fixes: https://tracker.ceph.com/issues/61378
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>