mon/AuthMonitor: add dump-keys and wipe-rotating-service-keys
`auth dump-keys` allows examining the key types for each entity and also the
rotating session keys. This lets us confirm key upgrades are done as expected.

`wipe-rotating-service-keys` clears out existing service keys so that we do not
need to wait for the rotating key expiration. It's a disruptive command however
as new clients will not be able to talk to service daemons without restarting
those daemons. Existing sessions will be unaffected until a normal rotation or
reconnect with the daemon.

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>