mon: provide emergency mechanism to use mon keyring
If they key is lost for the `mon.` credential, it's very inconvenient to get it
out of the "auth" database in the mon store. So, allow the operator to create a
new keyring for the mons and use it instead to get mons in quorum again.

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>
Change-Id: 46a7a5e3-a7e7-4a25-ad4e-710d5399c2a8
(cherry picked from commit bd6a8710958763d850592d409f1e3710836f80f0)