mgr/smb: Disable posix locking in share definition
The prerequisites for supporting durable handles[1] in Samba include
disabling the mapping of POSIX locks, as well as setting the `kernel
oplocks` and `kernel sharemodes` parameters to disabled. Currently
this configuration is hard‑coded, but in the future it could be made
conditional and combined with other settings to enable persistent
handles on continuously available shares.

[1] https://www.samba.org/samba/docs/current/man-html/smb.conf.5.html#DURABLEHANDLES

Conflicts:
	src/pybind/mgr/smb/handler.py

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>
Resolves: IBMCEPH-13335 (prereq)
(cherry picked from commit 05b125b2387736389e018c11a8219e07fabb4448)