auth: CryptoKey, dynamic usage keys
Make it possible to specify usage that was not initially declared
when the secret was set in CryptoKey.  As a prerequisite, simplify
the per-usage key storage structures so that is is less heavy-weight
and thread-safe.

Also, extend various encrypt and decrypt apis to enable passing
suage down to CryptoKey.

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Change-Id: e27aabcd-1511-4796-9461-c7bf7181dbd3
(cherry picked from commit 06cee3598a6ce2c14079021e5a32d485160a972e)